moving schedulers around
