+++ src/scheduler/NextTickScheduler.ts
@@ -1,4 +1,5 @@
 import Scheduler from './Scheduler';
+
 import {
 	NextScheduledAction,
 	ScheduledAction
