+++ src/scheduler/immediate.ts
@@ -1,0 +1,5 @@
+import Scheduler from './Scheduler';
+
+const immediate:Scheduler = new Scheduler();
+
+export default immediate;
