+++ src/scheduler/nextTick.ts
@@ -1,0 +1,5 @@
+import NextTickScheduler from './NextTickScheduler';
+
+const nextTick:NextTickScheduler = new NextTickScheduler();
+
+export default nextTick;
